major cleanup and reorganization with the full Run 2 fits included
