+++ configs/HHH/Run2_boosted_SR_pass_toy.json
@@ -21,7 +21,7 @@
           "TITLE":"t#bar{t}",
           "LOC":"PATH/FILE:HIST"
        },
-       "SIGNAME":{
+       "XToYHTo6B_MX-2500_MY-800":{
           "SYSTEMATICS":[
              "lumi"
           ],
@@ -33,51 +33,28 @@
         }
     },
     "REGIONS":{
-       "fail_boosted":{
+       "fail":{
           "PROCESSES":[
             "TTbar"
           ],
           "BINNING":"default"
        },
-       "pass_boosted":{
+       "pass":{
           "PROCESSES":[
             "TTbar",
-            "SIGNAME"
-          ],
-          "BINNING":"default"
-       },
-       "fail_semiboosted":{
-          "PROCESSES":[
-            "TTbar"
-          ],
-          "BINNING":"default"
-       },
-       "pass_semiboosted":{
-          "PROCESSES":[
-            "TTbar",
-            "SIGNAME"
+            "XToYHTo6B_MX-2500_MY-800"
           ],
           "BINNING":"default"
        }
     },
     "GLOBAL":{
-       "HIST":"mjj_vs_mjjj_SR_$region_nominal",
-       "HIST_UP":"mjj_vs_mjjj_SR_$region__$syst_up",
-       "HIST_DOWN":"mjj_vs_mjjj_SR_$region__$syst_down",
-       "HIST_ALT_UP":"mjj_vs_mjjj_SR_$region_$systUp",
-       "HIST_ALT_DOWN":"mjj_vs_mjjj_SR_$region_$systDown",
-       "PATH":"symlink2histograms_2017",
-       "FILE":"$process_Histograms.root",
-       "SIGNAME": [
-          "XToYHTo6B_MX-1000_MY-300", "XToYHTo6B_MX-1000_MY-600", "XToYHTo6B_MX-1000_MY-800",
-          "XToYHTo6B_MX-1200_MY-300", "XToYHTo6B_MX-1200_MY-600", "XToYHTo6B_MX-1200_MY-800", "XToYHTo6B_MX-1200_MY-1000",
-          "XToYHTo6B_MX-1600_MY-300", "XToYHTo6B_MX-1600_MY-600", "XToYHTo6B_MX-1600_MY-800", "XToYHTo6B_MX-1600_MY-1000", "XToYHTo6B_MX-1600_MY-1200", "XToYHTo6B_MX-1600_MY-1400",
-          "XToYHTo6B_MX-2000_MY-300", "XToYHTo6B_MX-2000_MY-600", "XToYHTo6B_MX-2000_MY-800", "XToYHTo6B_MX-2000_MY-1000", "XToYHTo6B_MX-2000_MY-1200", "XToYHTo6B_MX-2000_MY-1600", "XToYHTo6B_MX-2000_MY-1800",
-          "XToYHTo6B_MX-2500_MY-300", "XToYHTo6B_MX-2500_MY-600", "XToYHTo6B_MX-2500_MY-800", "XToYHTo6B_MX-2500_MY-1000", "XToYHTo6B_MX-2500_MY-1200", "XToYHTo6B_MX-2500_MY-1600", "XToYHTo6B_MX-2500_MY-2000", "XToYHTo6B_MX-2500_MY-2200",
-          "XToYHTo6B_MX-3000_MY-300", "XToYHTo6B_MX-3000_MY-600", "XToYHTo6B_MX-3000_MY-800", "XToYHTo6B_MX-3000_MY-1000", "XToYHTo6B_MX-3000_MY-1200", "XToYHTo6B_MX-3000_MY-1600", "XToYHTo6B_MX-3000_MY-2000", "XToYHTo6B_MX-3000_MY-2500", "XToYHTo6B_MX-3000_MY-2800",
-          "XToYHTo6B_MX-3500_MY-300", "XToYHTo6B_MX-3500_MY-600", "XToYHTo6B_MX-3500_MY-800", "XToYHTo6B_MX-3500_MY-1000", "XToYHTo6B_MX-3500_MY-1200", "XToYHTo6B_MX-3500_MY-1600", "XToYHTo6B_MX-3500_MY-2000", "XToYHTo6B_MX-3500_MY-2500", "XToYHTo6B_MX-3500_MY-2800",
-          "XToYHTo6B_MX-4000_MY-300", "XToYHTo6B_MX-4000_MY-600", "XToYHTo6B_MX-4000_MY-800", "XToYHTo6B_MX-4000_MY-1000", "XToYHTo6B_MX-4000_MY-1200", "XToYHTo6B_MX-4000_MY-1600", "XToYHTo6B_MX-4000_MY-2000", "XToYHTo6B_MX-4000_MY-2500", "XToYHTo6B_MX-4000_MY-2800"
-       ]
+       "HIST":"mjj_vs_mjjj_SR_$region_boosted_nominal",
+       "HIST_UP":"mjj_vs_mjjj_SR_$region_boosted__$syst_up",
+       "HIST_DOWN":"mjj_vs_mjjj_SR_$region_boosted__$syst_down",
+       "HIST_ALT_UP":"mjj_vs_mjjj_SR_$region_boosted_$systUp",
+       "HIST_ALT_DOWN":"mjj_vs_mjjj_SR_$region_boosted_$systDown",
+       "PATH":"symlink2histograms_Run2",
+       "FILE":"$process_Histograms.root"
     },
     "BINNING":{
        "default":{
