+++ configs/HHH/Run2_combined_SR_pass_toy_multiSignal.json
@@ -66,7 +66,7 @@
        "HIST_DOWN":"mjj_vs_mjjj_SR_$region__$syst_down",
        "HIST_ALT_UP":"mjj_vs_mjjj_SR_$region_$systUp",
        "HIST_ALT_DOWN":"mjj_vs_mjjj_SR_$region_$systDown",
-       "PATH":"symlink2histograms_2017",
+       "PATH":"symlink2histograms_Run2",
        "FILE":"$process_Histograms.root",
        "SIGNAME": [
           "XToYHTo6B_MX-1000_MY-300", "XToYHTo6B_MX-1000_MY-600", "XToYHTo6B_MX-1000_MY-800",
